cmdline: Use toDouble when parsing rate for scientific notation support

+++ main.cpp
@@ -25,7 +25,8 @@ int main(int argc, char *argv[])
 
     if (parser.isSet(rateOption)){
 	bool ok;
-	int rate = parser.value(rateOption).toInt(&ok);
+	// Use toDouble just for scientific notation support
+	int rate = parser.value(rateOption).toDouble(&ok);
 	if(!ok){
 	    fputs("ERROR: could not parse rate\n", stderr);
 	    return 1;
